Rename for consistency.
PermissionBinder→PermissionsBinder, data/paymentLink→data/paymentLinks, and data/subscription→data/subscriptions.
