+++ src/data/customers/CustomerHelper.ts
@@ -4,8 +4,8 @@ import Callback from '../../types/Callback';
 import Helper from '../Helper';
 import { PaymentData } from '../payments/data';
 import Payment from '../payments/Payment';
-import { SubscriptionData } from '../subscription/data';
-import Subscription from '../subscription/Subscription';
+import { SubscriptionData } from '../subscriptions/data';
+import Subscription from '../subscriptions/Subscription';
 import Customer, { CustomerData } from './Customer';
 import { MandateData } from './mandates/data';
 import Mandate from './mandates/Mandate';
